updated for local dynamic TDVP
